+++ xarray/core/merge.py
@@ -1,4 +1,4 @@
-from __future__ import absolute_import, division, print_function
+from collections import OrderedDict
 
 from typing import (
     Any, Dict, List, Mapping, Optional, Set, Tuple, TYPE_CHECKING, Union,
@@ -7,7 +7,6 @@ from typing import (
 import pandas as pd
 
 from .alignment import deep_align
-from .pycompat import OrderedDict, basestring
 from .utils import Frozen
 from .variable import (
     Variable, as_variable, assert_unique_multiindex_level_names)
@@ -543,7 +542,7 @@ def dataset_merge_method(dataset, other, overwrite_vars, compat, join):
     # method due for backwards compatibility
     # TODO: consider deprecating it?
 
-    if isinstance(overwrite_vars, basestring):
+    if isinstance(overwrite_vars, str):
         overwrite_vars = set([overwrite_vars])
     overwrite_vars = set(overwrite_vars)