+++ xarray/tests/test_utils.py
@@ -1,5 +1,4 @@
-from __future__ import absolute_import, division, print_function
-
+from collections import OrderedDict
 from datetime import datetime
 
 import numpy as np
@@ -9,7 +8,6 @@ import pytest
 import xarray as xr
 from xarray.coding.cftimeindex import CFTimeIndex
 from xarray.core import duck_array_ops, utils
-from xarray.core.pycompat import OrderedDict
 from xarray.core.utils import either_dict_or_kwargs
 from xarray.testing import assert_identical